Show tag value rather than name in tag list
The tag name (given in the JSON structure for the tagger) would be more
descriptive of course, but when we construct the multiword search view
from a CQP expression only the attribute value is available. We might
consider looking up the name in the JSON structure based on the
attribute name and the value.