Always produce attribute/value CQP expressions
i.e., expressions like [word="car" %c] instead of just "car", so that the search is case insensitive (we are still able to handle simple "car"-type expressions that are input manually in the regex view, though). Note that we should probably add an option to do case sensitive search (without using the regex view) at some point, and then we might reintroduce the ability to produce simple expressions.