Use uncontrolled text inputs for simple and cqp
This prevents the cursor from jumping to the end whenever we type something into a non-final position in the text box. Of course the text inputs need to be updated with a new value if we switch back to the view after having changed the query in another view. However, this is currently not a problem since only one view is visible at the time, meaning that the text input will be mounted, and hence the value set, each time we switch to it.