Specify keys for seqs of components.
Note that since several search results may have the same speaker IDs,
we cannot use those as keys. However, we can compute the hash code of
the entire result object (with ID, match, and pre- and post-match),
which should be unique (if two results are identical in all respects,
we probably don't want to show both of them anyway).