Implement 3-step search
Search is performed in three steps. In the first step, we create a new search object in the database and search for a single page of search results (as defined by the cut parameter sent by the client) so as to be able to display some results very quickly. In the second step, we retrieve the previously saved search object from the database and search for a larger number of search result pages (e.g. 20, but again defined by the cut parameter). In the final step, we do the same, but this time the cut parameter will be nil, causing the search to be unrestricted.